+++ history.md
@@ -42,7 +42,9 @@
 - 1.0.0-preview5
   - *added:* `Transient` services are also released on injector disposal (if they have not been disposed programmatically)
   - *added:* `ObjectExtensions.Act()` can be called against interfaces
-  - *added:* `ProxyGenerator` now supports `in` parameter modifier (introduced in C# 7.2)
+  - *added:* `ProxyGenerator` supports `in` parameter modifier (introduced in C# 7.2)
   - *fixed:* Ambiguous field naming in `ProxyGenerator`
   - *renamed:* `QueryOptionsAttribute` to `OptionsAttribute`
   - *renamed:* `QueryMode` to `QueryModes`
+- 1.0.0-preview6 (unreleased)
+  - *added:* `ProxyGenerator` supports `params` parameter modifier